requirements.txt
Django
pyjwt
djangorestframework
django-cors-headers
black
pytest
pytest-django